refactor: enhance assessment components with new data handling, UI updates, and performance reporting features
